% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictor.R
\name{CVLearnerPredictor}
\alias{CVLearnerPredictor}
\title{CVLearnerPredictor}
\description{
Wraps a mlr3 Learner into a \code{CVLearnerPredictor} object that can be used
with mcboost. Internally cross-validates predictions.
}
\section{Super class}{
\code{\link[mcboost:Predictor]{mcboost::Predictor}} -> \code{CVLearnerPredictor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pipeop}}{\code{\link{Learner}}\cr
mlr3pipelines PipeOp used for fitting residuals.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{is_fitted}}{\code{\link{logical}}\cr
Whether the Learner is trained}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CVLearnerPredictor$new()}}
\item \href{#method-fit_transform}{\code{CVLearnerPredictor$fit_transform()}}
\item \href{#method-predict}{\code{CVLearnerPredictor$predict()}}
\item \href{#method-clone}{\code{CVLearnerPredictor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mcboost" data-topic="Predictor" data-id="fit">}\href{../../mcboost/html/Predictor.html#method-fit}{\code{mcboost::Predictor$fit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a LearnerPredictor with internal cross-validation.
See \code{\link[mlr3pipelines:mlr_pipeops_learner_cv]{mlr3pipelines::PipeOpLearnerCV}} for more information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CVLearnerPredictor$new(learner, folds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learner}}{\code{\link{Learner}}\cr
Learner used for train/predict.}

\item{\code{folds}}{\code{\link{integer}}\cr
Number of folds to use for PipeOpLearnerCV.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
Fit the learner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CVLearnerPredictor$fit_transform(data, labels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}}\cr
Features.}

\item{\code{labels}}{\code{\link{numeric}}\cr
One-hot encoded labels (of same length as data).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict a dataset with leaner predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CVLearnerPredictor$predict(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}} \cr
Prediction data.}

\item{\code{...}}{\code{\link{any}} \cr
Not used, only for compatibility with other methods.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CVLearnerPredictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
