% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictor.R
\name{LearnerPredictor}
\alias{LearnerPredictor}
\title{LearnerPredictor
Wraps a mlr3 Learner into a \code{LearnerPredictor} object that can be used
with mcboost.}
\description{
LearnerPredictor
Wraps a mlr3 Learner into a \code{LearnerPredictor} object that can be used
with mcboost.

LearnerPredictor
Wraps a mlr3 Learner into a \code{LearnerPredictor} object that can be used
with mcboost.
}
\section{Super class}{
\code{\link[mcboost:Predictor]{mcboost::Predictor}} -> \code{LearnerPredictor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{learner}}{\code{\link{Learner}}\cr
mlr3 Learner used for fitting residuals.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{is_fitted}}{\code{\link{logical}}\cr
Whether the Learner is trained}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerPredictor$new()}}
\item \href{#method-fit}{\code{LearnerPredictor$fit()}}
\item \href{#method-predict}{\code{LearnerPredictor$predict()}}
\item \href{#method-clone}{\code{LearnerPredictor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a LearnerPredictor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerPredictor$new(learner)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learner}}{\code{\link{Learner}}\cr
Learner used for train/predict.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the learner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerPredictor$fit(data, labels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}}\cr
Features.}

\item{\code{labels}}{\code{\link{numeric}}\cr
One-hot encoded labels (of same length as data).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict a dataset with leaner predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerPredictor$predict(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}} \cr
Prediction data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerPredictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
