% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResidualFitters.R
\name{LearnerResidualFitter}
\alias{LearnerResidualFitter}
\alias{TreeResidualFitter}
\alias{RidgeResidualFitter}
\alias{CVTreeResidualFitter}
\alias{CVRidgeResidualFitter}
\title{ResidualFitter from a Learner}
\description{
ResidualFitter from a Learner

ResidualFitter from a Learner
}
\section{Functions}{
\itemize{
\item \code{TreeResidualFitter}: ResidualFitter based on rpart

\item \code{RidgeResidualFitter}: ResidualFitter based on glmnet

\item \code{CVTreeResidualFitter}: Cross-Validated residualFitter based on rpart

\item \code{CVRidgeResidualFitter}: Cross-Validated ResidualFitter based on glmnet
}}

\section{Super class}{
\code{\link[mcboost:ResidualFitter]{mcboost::ResidualFitter}} -> \code{LearnerResidualFitter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{learner}}{\code{\link{LearnerPredictor}}\cr
Learner used for fitting residuals.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerResidualFitter$new()}}
\item \href{#method-fit}{\code{LearnerResidualFitter$fit()}}
\item \href{#method-clone}{\code{LearnerResidualFitter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mcboost" data-topic="ResidualFitter" data-id="fit_to_resid">}\href{../../mcboost/html/ResidualFitter.html#method-fit_to_resid}{\code{mcboost::ResidualFitter$fit_to_resid()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Define a ResidualFitter from a Learner
Available instantiations: \code{\link{TreeResidualFitter}} (rpart) and
\code{\link{RidgeResidualFitter}} (glmnet).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerResidualFitter$new(learner)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learner}}{\code{\link{Learner}}\cr
Regression Learner to use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the learner and compute correlation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerResidualFitter$fit(data, resid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.frame}}\cr
Features to use.}

\item{\code{resid}}{\code{\link{numeric}}\cr
Target variable (residuals)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} with items\cr
\itemize{
\item \code{corr}: pseudo-correlation between residuals and learner prediction.
\item \code{l}: the trained learner.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerResidualFitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[mcboost:ResidualFitter]{mcboost::ResidualFitter}} -> \code{\link[mcboost:LearnerResidualFitter]{mcboost::LearnerResidualFitter}} -> \code{TreeResidualFitter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TreeResidualFitter$new()}}
\item \href{#method-clone}{\code{TreeResidualFitter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mcboost" data-topic="ResidualFitter" data-id="fit_to_resid">}\href{../../mcboost/html/ResidualFitter.html#method-fit_to_resid}{\code{mcboost::ResidualFitter$fit_to_resid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mcboost" data-topic="LearnerResidualFitter" data-id="fit">}\href{../../mcboost/html/LearnerResidualFitter.html#method-fit}{\code{mcboost::LearnerResidualFitter$fit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Define a ResidualFitter from a rpart learner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeResidualFitter$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeResidualFitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[mcboost:ResidualFitter]{mcboost::ResidualFitter}} -> \code{\link[mcboost:LearnerResidualFitter]{mcboost::LearnerResidualFitter}} -> \code{RidgeResidualFitter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RidgeResidualFitter$new()}}
\item \href{#method-clone}{\code{RidgeResidualFitter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mcboost" data-topic="ResidualFitter" data-id="fit_to_resid">}\href{../../mcboost/html/ResidualFitter.html#method-fit_to_resid}{\code{mcboost::ResidualFitter$fit_to_resid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mcboost" data-topic="LearnerResidualFitter" data-id="fit">}\href{../../mcboost/html/LearnerResidualFitter.html#method-fit}{\code{mcboost::LearnerResidualFitter$fit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Define a ResidualFitter from a glmnet learner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RidgeResidualFitter$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RidgeResidualFitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[mcboost:ResidualFitter]{mcboost::ResidualFitter}} -> \code{\link[mcboost:CVLearnerResidualFitter]{mcboost::CVLearnerResidualFitter}} -> \code{CVTreeResidualFitter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CVTreeResidualFitter$new()}}
\item \href{#method-clone}{\code{CVTreeResidualFitter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mcboost" data-topic="ResidualFitter" data-id="fit_to_resid">}\href{../../mcboost/html/ResidualFitter.html#method-fit_to_resid}{\code{mcboost::ResidualFitter$fit_to_resid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mcboost" data-topic="CVLearnerResidualFitter" data-id="fit">}\href{../../mcboost/html/CVLearnerResidualFitter.html#method-fit}{\code{mcboost::CVLearnerResidualFitter$fit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Define a ResidualFitter from a rpart learner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CVTreeResidualFitter$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CVTreeResidualFitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[mcboost:ResidualFitter]{mcboost::ResidualFitter}} -> \code{\link[mcboost:CVLearnerResidualFitter]{mcboost::CVLearnerResidualFitter}} -> \code{CVRidgeResidualFitter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CVRidgeResidualFitter$new()}}
\item \href{#method-clone}{\code{CVRidgeResidualFitter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mcboost" data-topic="ResidualFitter" data-id="fit_to_resid">}\href{../../mcboost/html/ResidualFitter.html#method-fit_to_resid}{\code{mcboost::ResidualFitter$fit_to_resid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mcboost" data-topic="CVLearnerResidualFitter" data-id="fit">}\href{../../mcboost/html/CVLearnerResidualFitter.html#method-fit}{\code{mcboost::CVLearnerResidualFitter$fit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Define a ResidualFitter from a glmnet learner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CVRidgeResidualFitter$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CVRidgeResidualFitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
