% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_survival_curve}
\alias{make_survival_curve}
\title{Make every row monotonically decreasing in order to obtain the survival property.
Additionally, many predicitions need 1 as a first value and 0 as a last value.
(e.g. \code{PredictionSurv} needs this attribute.)}
\usage{
make_survival_curve(prediction)
}
\arguments{
\item{prediction}{\code{\link{data.table}}
Data.table with predictions. Every row is survival probability for the corresponding time.
Every column corresponds to  a specific time point.}
}
\value{
\code{\link{data.table}}
}
\description{
Make every row monotonically decreasing in order to obtain the survival property.
Additionally, many predicitions need 1 as a first value and 0 as a last value.
(e.g. \code{PredictionSurv} needs this attribute.)
}
