% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mlr3_init_predictor}
\alias{mlr3_init_predictor}
\title{Create an initial predictor from a mlr3 learner}
\usage{
mlr3_init_predictor(learner)
}
\arguments{
\item{learner}{\code{\link[mlr3:Learner]{mlr3::Learner}}
A trained learner used for initialization.}
}
\description{
Create an initial predictor from a mlr3 learner
}
