% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpMCBoost.R
\name{ppl_mcboost}
\alias{ppl_mcboost}
\title{Multi-calibration pipeline}
\usage{
ppl_mcboost(lrn = lrn("classif.featureless"))
}
\arguments{
\item{lrn}{\code{\link{Learner}}\cr
Initial learner. Internally wrapped into a \code{PipeOpLearnerCV}
with \code{resampling.method = "insample"} as a default.
All parameters can be adjusted through the resulting \verb{Graph's param_set. Defaults to }lrn("classif.featureless")\verb{. Note: An initial predictor can also be supplied via the }init_predictor` parameter.}
}
\description{
Wraps MCBoost in a Pipeline to be used with \code{mlr3pipelines}.
For now this assumes training on the same dataset that is later used
for multi-calibration.
}
