% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbRange.R
\name{ProbRange}
\alias{ProbRange}
\title{Range of Probabilities}
\description{
Range of format [lower; upper).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lower}}{\code{\link{numeric}} \cr
Lower bound of the ProbRange.}

\item{\code{upper}}{\code{\link{numeric}} \cr
upper bound of the ProbRange.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProbRange$new()}}
\item \href{#method-is_equal}{\code{ProbRange$is_equal()}}
\item \href{#method-is_not_equal}{\code{ProbRange$is_not_equal()}}
\item \href{#method-in_range_mask}{\code{ProbRange$in_range_mask()}}
\item \href{#method-print}{\code{ProbRange$print()}}
\item \href{#method-clone}{\code{ProbRange$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a Probability Range
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbRange$new(lower = -Inf, upper = Inf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lower}}{\code{\link{numeric}}\cr
Lower bound of the ProbRange.}

\item{\code{upper}}{\code{\link{numeric}}\cr
Upper bound of the ProbRange.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_equal"></a>}}
\if{latex}{\out{\hypertarget{method-is_equal}{}}}
\subsection{Method \code{is_equal()}}{
Compare with 'other' Probability Range regarding equality
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbRange$is_equal(other)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{other}}{\code{\link{ProbRange}}\cr
ProbRange to compare to.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical, whether ProbRanges are equal.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_not_equal"></a>}}
\if{latex}{\out{\hypertarget{method-is_not_equal}{}}}
\subsection{Method \code{is_not_equal()}}{
Compare with 'other' Probability Range regarding in-equality
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbRange$is_not_equal(other)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{other}}{\code{\link{ProbRange}}\cr
ProbRange to compare to.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical, whether ProbRanges are in-equal.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-in_range_mask"></a>}}
\if{latex}{\out{\hypertarget{method-in_range_mask}{}}}
\subsection{Method \code{in_range_mask()}}{
Check whether elements of an array are in the ProbRange.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbRange$in_range_mask(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link{numeric}}\cr
Array of probabilities}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical array, whether elements are in ProbRange or not.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer for ProbRange
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbRange$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbRange$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
