% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResidualFitters.R
\name{ResidualFitter}
\alias{ResidualFitter}
\title{ResidualFitter Abstract Base Class}
\description{
ResidualFitter Abstract Base Class

ResidualFitter Abstract Base Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fit_to_resid}{\code{ResidualFitter$fit_to_resid()}}
\item \href{#method-fit}{\code{ResidualFitter$fit()}}
\item \href{#method-clone}{\code{ResidualFitter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit_to_resid"></a>}}
\if{latex}{\out{\hypertarget{method-fit_to_resid}{}}}
\subsection{Method \code{fit_to_resid()}}{
Fit to residuals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResidualFitter$fit_to_resid(data, resid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}}\cr
Features.}

\item{\code{data}}{\code{\link{data.table}}\cr
Features.}

\item{\code{resid}}{\code{\link{numeric}}\cr
Residuals (of same length as data).}

\item{\code{resid}}{\code{\link{numeric}}\cr
Residuals (of same length as data).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit to residuals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResidualFitter$fit(data, resid, mask)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}}\cr
Features.}

\item{\code{data}}{\code{\link{data.table}}\cr
Features.}

\item{\code{resid}}{\code{\link{numeric}}\cr
Residuals (of same length as data).}

\item{\code{resid}}{\code{\link{numeric}}\cr
Residuals (of same length as data).}

\item{\code{mask}}{\code{\link{integer}}\cr
Mask applied to the data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResidualFitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
