% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResidualFitters.R
\name{SubgroupFitter}
\alias{SubgroupFitter}
\title{Static ResidualFitter based on Subgroups}
\description{
Static ResidualFitter based on Subgroups

Static ResidualFitter based on Subgroups
}
\section{Super class}{
\code{\link[mcboost:ResidualFitter]{mcboost::ResidualFitter}} -> \code{SubgroupFitter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{subgroup_masks}}{\code{\link{list}} \cr
List of subgroup masks.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubgroupFitter$new()}}
\item \href{#method-fit}{\code{SubgroupFitter$fit()}}
\item \href{#method-clone}{\code{SubgroupFitter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mcboost" data-topic="ResidualFitter" data-id="fit_to_resid">}\href{../../mcboost/html/ResidualFitter.html#method-fit_to_resid}{\code{mcboost::ResidualFitter$fit_to_resid()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize SubgroupFitter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubgroupFitter$new(subgroup_masks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subgroup_masks}}{\code{\link{list}} \cr
List of subgroup masks. Subgroup masks are list(s) of integer masks,
each with the same length as data to be fitted on.
They allow defining sub-groups of the data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the learner and compute correlation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubgroupFitter$fit(data, resid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.frame}}\cr
Features to use.}

\item{\code{resid}}{\code{\link{numeric}}\cr
Target variable (residuals)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubgroupFitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
