% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictor.R
\name{SubgroupModel}
\alias{SubgroupModel}
\title{SubgroupModel}
\description{
SubgroupModel

SubgroupModel
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{subgroup_masks}}{\code{\link{list}} \cr
List of subgroup masks.}

\item{\code{subgroup_preds}}{\code{\link{list}} \cr
List of subgroup predictions after fitting.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubgroupModel$new()}}
\item \href{#method-fit}{\code{SubgroupModel$fit()}}
\item \href{#method-predict}{\code{SubgroupModel$predict()}}
\item \href{#method-clone}{\code{SubgroupModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a SubpopPredictor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubgroupModel$new(subgroup_masks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subgroup_masks}}{\code{\link{list}} \cr
List of subgroup masks.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the predictor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubgroupModel$fit(data, labels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}}\cr
Features.}

\item{\code{labels}}{\code{\link{numeric}}\cr
One-hot encoded labels (of same length as data).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict a dataset with sub-population predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubgroupModel$predict(data, subgroup_masks = NULL, partition_mask = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}} \cr
Prediction data.}

\item{\code{subgroup_masks}}{\code{\link{list}} \cr
List of subgroup masks for the data.}

\item{\code{partition_mask}}{\code{\link{integer}} \cr
Mask defined by partitions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubgroupModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
