% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictor.R
\name{ConstantPredictor}
\alias{ConstantPredictor}
\title{ConstantPredictor}
\description{
ConstantPredictor

ConstantPredictor
}
\section{Super class}{
\code{\link[mcboost:Predictor]{mcboost::Predictor}} -> \code{ConstantPredictor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{constant}}{\code{\link{numeric}}\cr
mlr3 Constant to predict with.}

\item{\code{is_fitted}}{\code{\link{logical}}\cr
Whether the model is fitted.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConstantPredictor$new()}}
\item \href{#method-fit}{\code{ConstantPredictor$fit()}}
\item \href{#method-predict}{\code{ConstantPredictor$predict()}}
\item \href{#method-clone}{\code{ConstantPredictor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a ConstantPredictor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstantPredictor$new(constant = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{constant}}{\code{\link{numeric}}\cr
Constant to predict with.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the constant predictor.
Does nothing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstantPredictor$fit(data, labels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}}\cr
Features.}

\item{\code{labels}}{\code{\link{numeric}}\cr
One-hot encoded labels (of same length as data).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict a dataset with constant predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstantPredictor$predict(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}} \cr
Prediction data.}

\item{\code{...}}{\code{\link{any}} \cr
Not used, only for compatibility with other methods.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstantPredictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
