% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResidualFitters.R
\name{CVLearnerResidualFitter}
\alias{CVLearnerResidualFitter}
\title{ResidualFitter from a Learner}
\description{
ResidualFitter from a Learner

ResidualFitter from a Learner
}
\section{Super class}{
\code{\link[mcboost:ResidualFitter]{mcboost::ResidualFitter}} -> \code{CVLearnerResidualFitter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{learner}}{\code{\link{LearnerPredictor}}\cr
Learner used for fitting residuals.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CVLearnerResidualFitter$new()}}
\item \href{#method-fit}{\code{CVLearnerResidualFitter$fit()}}
\item \href{#method-clone}{\code{CVLearnerResidualFitter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mcboost" data-topic="ResidualFitter" data-id="fit_to_resid">}\href{../../mcboost/html/ResidualFitter.html#method-fit_to_resid}{\code{mcboost::ResidualFitter$fit_to_resid()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Define a ResidualFitter from a Learner
Available instantiations: \code{\link{TreeResidualFitter}} (rpart) and
\code{\link{RidgeResidualFitter}} (glmnet).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CVLearnerResidualFitter$new(learner)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learner}}{\code{\link{Learner}}\cr
Regression Learner to use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the cv-learner and compute correlation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CVLearnerResidualFitter$fit(data, resid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.frame}}\cr
Features to use.}

\item{\code{resid}}{\code{\link{numeric}}\cr
Target variable (residuals)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} with items\cr
\itemize{
\item \code{corr}: pseudo-correlation between residuals and learner prediction.
\item \code{l}: the trained learner.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CVLearnerResidualFitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
