% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mcboost-package}
\alias{mcboost}
\alias{mcboost-package}
\title{mcboost: Implementes Multi-Accuracy Boosting (Kim et al., 2018).}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\references{
https://arxiv.org/pdf/1805.12317.pdf (Kim et al. 2018)
}
