% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCBoost.R
\name{MCBoost}
\alias{MCBoost}
\title{Multi-Accuracy Boosting for R}
\description{
For more details, please refer to https://arxiv.org/pdf/1805.12317.pdf (Kim et al. 2018).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{max_iter}}{\code{\link{integer}} \cr
The maximum number of iterations of the multicalibration/multiaccuracy method.}

\item{\code{alpha}}{\code{\link{numeric}} \cr
Accuracy parameter that determines the stopping condition.}

\item{\code{eta}}{\code{\link{numeric}} \cr
Parameter for multiplicative weight update (step size).}

\item{\code{num_buckets}}{\code{\link{integer}} \cr
The number of buckets to split into.}

\item{\code{bucket_strategy}}{\code{\link{character}} \cr
Currently onle supports "simple", even split along probabilities.
Only relevant for num_buckets > 1.}

\item{\code{rebucket}}{\code{\link{logical}} \cr
Should buckets be re-done at each iteration?}

\item{\code{partition}}{\code{\link{logical}} \cr
True/False flag for whether to split up predictions by their "partition"
(e.g., predictions less than
0.5 and predictions greater than 0.5)}

\item{\code{multiplicative}}{\code{\link{logical}} \cr
Specifies the strategy for updating the weights (multiplicative weight vs additive).}

\item{\code{subpop_fitter}}{\code{\link{ResidualFitter}} \cr
Specifies the type of model used to fit the
residual fxn ('TreeResidualFitter' or 'RidgeResidualFitter' (default)).}

\item{\code{predictor}}{\code{\link{function}} \cr
Initial redictor function.}

\item{\code{iter_models}}{\code{\link{list}} \cr
Cummulative list of fitted models.}

\item{\code{iter_partitions}}{\code{\link{list}} \cr
Cummulative list of data partitions for models.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MCBoost$new()}}
\item \href{#method-multicalibrate}{\code{MCBoost$multicalibrate()}}
\item \href{#method-predict_probs}{\code{MCBoost$predict_probs()}}
\item \href{#method-clone}{\code{MCBoost$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a multi-calibration instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCBoost$new(
  max_iter = 5,
  alpha = 1e-04,
  eta = 1,
  partition = FALSE,
  num_buckets = 2,
  bucket_strategy = "simple",
  rebucket = FALSE,
  multiplicative = FALSE,
  subpop_fitter = NULL,
  subpops = NULL,
  default_model_class = ConstantPredictor,
  init_predictor = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_iter}}{\code{\link{integer}} \cr
The maximum number of iterations of the multicalibration/multiaccuracy method.}

\item{\code{alpha}}{\code{\link{numeric}} \cr
Accuracy parameter that determines the stopping condition.}

\item{\code{eta}}{\code{\link{numeric}} \cr
Parameter for multiplicative weight update (step size).}

\item{\code{partition}}{\code{\link{logical}} \cr
True/False flag for whether to split up predictions by their "partition"
(e.g., predictions less than
0.5 and predictions greater than 0.5)}

\item{\code{num_buckets}}{\code{\link{integer}} \cr
The number of buckets to split into.}

\item{\code{bucket_strategy}}{\code{\link{character}} \cr
Currently onle supports "simple", even split along probabilities.
Only relevant for num_buckets > 1.}

\item{\code{rebucket}}{\code{\link{logical}} \cr
Should buckets be re-done at each iteration?}

\item{\code{multiplicative}}{\code{\link{logical}} \cr
Specifies the strategy for updating the weights (multiplicative weight vs additive)}

\item{\code{subpop_fitter}}{\code{\link{ResidualFitter}} \cr
Specifies the type of model used to fit the
residual \link{'TreeResidualFitter'} or \link{'RidgeResidualFitter'} (default)).}

\item{\code{subpops}}{\code{\link{list}} \cr
Specifies a collection of characteristic attributes
and the values they take defining the S in subpops
e.g. C = list(age = c('20-29','30-39','40+'), nJobs = c(0,1,2,'3+'), ,..).}

\item{\code{default_model_class}}{\code{\link{Predictor}} \cr
The class of the model that should be used
as the MCBoost's default predictor model.}

\item{\code{init_predictor}}{\code{\link{function}} \cr
The initial predictor function to use (i.e., if
the user has a pretrained model).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multicalibrate"></a>}}
\if{latex}{\out{\hypertarget{method-multicalibrate}{}}}
\subsection{Method \code{multicalibrate()}}{
Run multicalibration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCBoost$multicalibrate(data, labels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}}\cr
Features.}

\item{\code{labels}}{\code{\link{numeric}}\cr
One-hot encoded labels (of same length as data).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_probs"></a>}}
\if{latex}{\out{\hypertarget{method-predict_probs}{}}}
\subsection{Method \code{predict_probs()}}{
Predict a dataset with multi-calibrated predictions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCBoost$predict_probs(x, t = Inf, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link{data.table}} \cr
Prediction data.}

\item{\code{t}}{\code{\link{integer}} \cr
Number of multi-calibration steps to predict. Default: \code{Inf} (all).}

\item{\code{...}}{\code{\link{any}} \cr
Params passed on to other methods. Currently not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Numeric vector of multi-calibrated predictions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCBoost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
