% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{one_hot}
\alias{one_hot}
\title{One-hot encode a factor variable}
\usage{
one_hot(labels)
}
\arguments{
\item{labels}{\code{\link{factor}}\cr
Factor to encode.}
}
\value{
\code{\link{integer}}\cr
Integer vector of encoded labels.
}
\description{
One-hot encode a factor variable
}
\examples{
  one_hot(factor(c("a", "b", "a")))
}
