% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpMCBoost.R
\name{ppl_mcboost}
\alias{ppl_mcboost}
\title{Multi-calibration pipeline}
\usage{
ppl_mcboost(lrn = lrn("classif.rpart"))
}
\arguments{
\item{lrn}{\code{\link{Learner}}\cr
Initial learner. Internally wrapped into a \code{PipeOpLearnerCV}
with \code{resampling.method = "insample"} as a default.
All parameters can be adjusted through the resulting `Graph's param_set.}
}
\description{
Wraps MCBoost in a Pipeline to be used with \code{mlr3pipelines}.
For now this assumes training on the same dataset that is later used
for multi-calibration.
}
