% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResidualFitters.R
\name{SubpopFitter}
\alias{SubpopFitter}
\title{Static ResidualFitter based on Subpopulations}
\description{
Static ResidualFitter based on Subpopulations

Static ResidualFitter based on Subpopulations
}
\seealso{
Other SubPopFitters: 
\code{\link{CVLearnerResidualFitter}},
\code{\link{LearnerResidualFitter}},
\code{\link{ResidualFitter}},
\code{\link{SubgroupFitter}}
}
\concept{SubPopFitters}
\section{Super class}{
\code{\link[mcboost:ResidualFitter]{mcboost::ResidualFitter}} -> \code{SubpopFitter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{subgroup_masks}}{\code{\link{list}} \cr
List of subgroup masks.}

\item{\code{subpops}}{\code{\link{list}} \cr
List of subpopulation indicators.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubpopFitter$new()}}
\item \href{#method-fit}{\code{SubpopFitter$fit()}}
\item \href{#method-clone}{\code{SubpopFitter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mcboost" data-topic="ResidualFitter" data-id="fit_to_resid">}\href{../../mcboost/html/ResidualFitter.html#method-fit_to_resid}{\code{mcboost::ResidualFitter$fit_to_resid()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize SubPopFilter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubpopFitter$new(subpops)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subpops}}{\code{\link{list}} \cr
Specifies a collection of characteristic attributes
and the values they take defining the S in subpops
e.g. C = list(age = c('20-29','30-39','40+'), nJobs = c(0,1,2,'3+'), ,..).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the learner and compute correlation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubpopFitter$fit(data, resid, mask)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}}\cr
Features.}

\item{\code{resid}}{\code{\link{numeric}}\cr
Residuals (of same length as data).}

\item{\code{mask}}{\code{\link{integer}}\cr
Mask applied to the data. Only used for \code{SubgroupFitter}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} with items\cr
\itemize{
\item \code{corr}: pseudo-correlation between residuals and learner prediction.
\item \code{l}: the trained learner.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubpopFitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
