% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictor.R
\name{SubpopPredictor}
\alias{SubpopPredictor}
\title{SubpopPredictor}
\description{
SubpopPredictor

SubpopPredictor
}
\section{Super class}{
\code{\link[mcboost:Predictor]{mcboost::Predictor}} -> \code{SubpopPredictor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{subpop}}{\code{\link{function}} \cr
A \code{\link{function}} that evaluates to binary for each row in a dataset.
Defines a sub-population.}

\item{\code{value}}{\code{\link{numeric}} \cr
A correlation value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubpopPredictor$new()}}
\item \href{#method-fit}{\code{SubpopPredictor$fit()}}
\item \href{#method-predict}{\code{SubpopPredictor$predict()}}
\item \href{#method-clone}{\code{SubpopPredictor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a SubpopPredictor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubpopPredictor$new(subpop, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subpop}}{\code{\link{character}}|\code{\link{function}} \cr
Either a \code{\link{function}}, that yields a binary value for each
row in a dataset, or a \code{\link{character}} string referring to a
feature column, that defines a sub-population.}

\item{\code{value}}{\code{\link{numeric}} \cr
Correlation value for the given subpop.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the predictor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubpopPredictor$fit(data, labels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}}\cr
Features.}

\item{\code{labels}}{\code{\link{numeric}}\cr
One-hot encoded labels (of same length as data).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict a dataset with sub-population predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubpopPredictor$predict(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{\link{data.table}} \cr
Prediction data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubpopPredictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
